requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-cors-headers==3.4.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
mercadopago==1.1.1
psycopg2==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
